my-pr/requirements.txt
Django==2.0.7
django-filter==2.0.0
django-filters==0.2.1
django-jinja==2.4.1
django-search-views==0.3.7
django-tinymce==2.7.0
django-treebeard==4.3
django-widget-tweaks==1.4.2
Jinja2==2.10
MarkupSafe==1.0
Pillow==5.2.0
pytz==2018.5
